refactor: extract socnav occupancy helpers
