+++ docs/context/README.md
@@ -106,6 +106,8 @@ knowledge, not every transient iteration detail.
   [issue_1584_socnav_unavailable_row_policy.md](issue_1584_socnav_unavailable_row_policy.md)
 * Issue #1583 high-risk root path boundaries:
   [issue_1583_high_risk_root_boundaries.md](issue_1583_high_risk_root_boundaries.md)
+* Issue #1634 SocNav module split plan:
+  [issue_1634_socnav_split_plan.md](issue_1634_socnav_split_plan.md)
 * Issues #1598/#1599 Root Compatibility Decisions (2026-05-28):
   [issue_1598_1599_root_compatibility_decisions.md](issue_1598_1599_root_compatibility_decisions.md)
 * Issue #1504 ego-conditioned feature contract:
